// src/pages/Home/Home.types.ts
export interface IGameStatus {
  finished: boolean
  success: boolean
}

export const congratulatoryText = 'Поздравляем вас с Днём Рождения, желаем счастья, здоровья, любви, радости, веселья, успеха, удачи, улыбок и хорошего настроения'